Move some common fields from Select and SelectUnion to Query
