+++ h2/src/main/org/h2/command/dml/Query.java
@@ -10,6 +10,7 @@ import java.util.HashSet;
 import java.util.List;
 
 import org.h2.api.ErrorCode;
+import org.h2.command.CommandInterface;
 import org.h2.command.Prepared;
 import org.h2.engine.Database;
 import org.h2.engine.Mode.ModeEnum;
@@ -40,39 +41,55 @@ import org.h2.value.ValueNull;
 public abstract class Query extends Prepared {
 
     /**
+     * The column list, including invisible expressions such as order by expressions.
+     */
+    ArrayList<Expression> expressions;
+
+    /**
+     * Array of expressions.
+     *
+     * @see #expressions
+     */
+    Expression[] expressionArray;
+
+    ArrayList<SelectOrderBy> orderList;
+
+    SortOrder sort;
+
+    /**
      * The limit expression as specified in the LIMIT or TOP clause.
      */
-    protected Expression limitExpr;
+    Expression limitExpr;
 
     /**
      * Whether limit expression specifies percentage of rows.
      */
-    protected boolean fetchPercent;
+    boolean fetchPercent;
 
     /**
      * Whether tied rows should be included in result too.
      */
-    protected boolean withTies;
+    boolean withTies;
 
     /**
      * The offset expression as specified in the LIMIT .. OFFSET clause.
      */
-    protected Expression offsetExpr;
+    Expression offsetExpr;
 
     /**
      * The sample size expression as specified in the SAMPLE_SIZE clause.
      */
-    protected Expression sampleSizeExpr;
+    Expression sampleSizeExpr;
 
     /**
      * Whether the result must only contain distinct rows.
      */
-    protected boolean distinct;
+    boolean distinct;
 
     /**
      * Whether the result needs to support random access.
      */
-    protected boolean randomAccessResult;
+    boolean randomAccessResult;
 
     private boolean noCache;
     private int lastLimit;
@@ -144,7 +161,9 @@ public abstract class Query extends Prepared {
      *
      * @return the list of expressions
      */
-    public abstract ArrayList<Expression> getExpressions();
+    public ArrayList<Expression> getExpressions() {
+        return expressions;
+    }
 
     /**
      * Calculate the cost to execute this query.
@@ -178,14 +197,18 @@ public abstract class Query extends Prepared {
      *
      * @param order the order by list
      */
-    public abstract void setOrder(ArrayList<SelectOrderBy> order);
+    public void setOrder(ArrayList<SelectOrderBy> order) {
+        orderList = order;
+    }
 
     /**
      * Whether the query has an order.
      *
      * @return true if it has
      */
-    public abstract boolean hasOrder();
+    public boolean hasOrder() {
+        return orderList != null || sort != null;
+    }
 
     /**
      * Set the 'for update' flag.
@@ -598,6 +621,11 @@ public abstract class Query extends Prepared {
             sortType[i] = type;
         }
         return new SortOrder(session.getDatabase(), index, sortType, orderList);
+    }
+
+    @Override
+    public int getType() {
+        return CommandInterface.SELECT;
     }
 
     public void setOffset(Expression offset) {
